+++ src/Gitea/Gitea_API.php
@@ -164,7 +164,6 @@ class Gitea_API extends API implements API_Interface {
 		self::$method       = 'download_link';
 		$download_link_base = $this->get_api_url( '/repos/:owner/:repo/archive/', true );
 		$endpoint           = '';
-		$cache              = $this->get_repo_cache( $this->type->slug ?? false, false );
 
 		// Release asset.
 		if ( $this->use_release_asset( $branch_switch ) ) {
@@ -189,14 +188,8 @@ class Gitea_API extends API implements API_Interface {
 				}
 			}
 
-			if ( empty( $cache['release_asset_download'] ) ) {
-				$this->set_repo_cache( 'release_asset_download', $release_asset );
-			}
-			if ( ! empty( $cache['release_asset_download'] ) ) {
-				return $cache['release_asset_download'];
-			}
-
-			return $this->get_release_asset_redirect( $release_asset, true );
+			$this->set_repo_cache( 'release_asset_download', $release_asset );
+			return $release_asset;
 		}
 
 		/*