+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* show/hide settings fields based on OAuth setting
 
 #### 2.6.0 / 2026-05-31 🎂
 * add release assets
@@ -8,7 +9,6 @@
 * move download base outside of loop
 * add `gu_dev_release_asset` filter
 * use local variables
-* show/hide settings fields based on OAuth setting
 
 #### 2.5.0 / 2025-08-04
 * add error checking to `parse_contents_response()`