+++ CHANGES.md
@@ -1,6 +1,7 @@
 [unreleased]
 * show/hide settings fields based on OAuth setting
 * add note about tokens stored in plaintext
+* no longer using release asset redirect
 
 #### 2.6.0 / 2026-05-31 🎂
 * add release assets